Move OneShotTask in itw own class
