+++ merlin-driver/src/test/java/gov/nasa/jpl/aerie/merlin/driver/SimulationDuplicationTest.java
@@ -112,7 +112,7 @@ public class SimulationDuplicationTest {
       final List<Duration> desiredCheckpoints,
       final CachedEngineStore engineStore
   ) {
-    return CheckpointSimulationDriver.computeResults(CheckpointSimulationDriver.simulateWithCheckpoints(
+    return SimulationResultsComputerInputs.computeResults(CheckpointSimulationDriver.simulateWithCheckpoints(
         TestMissionModel.missionModel(),
         Map.of(),
         Instant.EPOCH,
@@ -125,7 +125,8 @@ public class SimulationDuplicationTest {
         CheckpointSimulationDriver.desiredCheckpoints(desiredCheckpoints),
         CheckpointSimulationDriver.noCondition(),
         engineStore,
-        mockConfiguration()
+        mockConfiguration(),
+        false
         ));
   }
 }
