+++ merlin-driver/src/test/java/gov/nasa/jpl/aerie/merlin/driver/TestMissionModel.java
@@ -3,13 +3,11 @@ package gov.nasa.jpl.aerie.merlin.driver;
 import gov.nasa.jpl.aerie.merlin.driver.timeline.LiveCells;
 import gov.nasa.jpl.aerie.merlin.driver.timeline.TemporalEventSource;
 import gov.nasa.jpl.aerie.merlin.protocol.driver.Initializer;
-import gov.nasa.jpl.aerie.merlin.protocol.driver.Scheduler;
 import gov.nasa.jpl.aerie.merlin.protocol.driver.Topic;
 import gov.nasa.jpl.aerie.merlin.protocol.model.DirectiveType;
 import gov.nasa.jpl.aerie.merlin.protocol.model.InputType;
 import gov.nasa.jpl.aerie.merlin.protocol.model.ModelType;
 import gov.nasa.jpl.aerie.merlin.protocol.model.OutputType;
-import gov.nasa.jpl.aerie.merlin.protocol.model.Task;
 import gov.nasa.jpl.aerie.merlin.protocol.model.TaskFactory;
 import gov.nasa.jpl.aerie.merlin.protocol.types.Duration;
 import gov.nasa.jpl.aerie.merlin.protocol.types.InSpan;
@@ -24,8 +22,7 @@ import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Executor;
-import java.util.function.Function;
+
 public class TestMissionModel {
   private final static Duration oneMinute = Duration.of(60, Duration.SECONDS);
   private static final Topic<Object> delayedActivityDirectiveInputTopic = new Topic<>();
@@ -52,9 +49,9 @@ public class TestMissionModel {
 
     @Override
     public TaskFactory<Object> getTaskFactory(final Object o, final Object o2) {
-      return executor -> oneShotTask($ -> {
+      return executor -> new OneStepTask<>($ -> {
         $.emit(this, delayedActivityDirectiveInputTopic);
-        return TaskStatus.delayed(oneMinute, oneShotTask($$ -> {
+        return TaskStatus.delayed(oneMinute, new OneStepTask<>($$ -> {
           $$.emit(Unit.UNIT, delayedActivityDirectiveOutputTopic);
           return TaskStatus.completed(Unit.UNIT);
         }));
@@ -77,18 +74,18 @@ public class TestMissionModel {
 
     @Override
     public TaskFactory<Object> getTaskFactory(final Object o, final Object o2) {
-      return executor -> oneShotTask(scheduler -> {
+      return executor -> new OneStepTask<>(scheduler -> {
         scheduler.emit(this, decomposingActivityDirectiveInputTopic);
         return TaskStatus.delayed(
             Duration.ZERO,
-            oneShotTask($ -> {
+            new OneStepTask<>($ -> {
               try {
                 $.spawn(InSpan.Fresh, delayedActivityDirective.getTaskFactory(null, null));
               } catch (final InstantiationException ex) {
                 throw new Error("Unexpected state: activity instantiation of DelayedActivityDirective failed with: %s".formatted(
                     ex.toString()));
               }
-              return TaskStatus.delayed(Duration.of(120, Duration.SECOND), oneShotTask($$ -> {
+              return TaskStatus.delayed(Duration.of(120, Duration.SECOND), new OneStepTask<>($$ -> {
                 try {
                   $$.spawn(InSpan.Fresh, delayedActivityDirective.getTaskFactory(null, null));
                 } catch (final InstantiationException ex) {
@@ -103,20 +100,6 @@ public class TestMissionModel {
       });
     }
   };
-
-  private static <T> Task<T> oneShotTask(Function<Scheduler, TaskStatus<T>> f) {
-    return new Task<>() {
-      @Override
-      public TaskStatus<T> step(final Scheduler scheduler) {
-        return f.apply(scheduler);
-      }
-
-      @Override
-      public Task<T> duplicate(Executor executor) {
-        return this;
-      }
-    };
-  }
 
   private static final InputType<Object> testModelInputType = new InputType<>() {
     @Override
