+++ scheduler-worker/src/test/java/gov/nasa/jpl/aerie/scheduler/worker/services/MockMerlinService.java
@@ -151,7 +151,7 @@ class MockMerlinService implements MerlinService.OwnerRole {
   }
 
   @Override
-  public Optional<SimulationResults> getSimulationResults(final PlanMetadata planMetadata)
+  public Optional<Pair<SimulationResults, DatasetId>> getSimulationResults(final PlanMetadata planMetadata)
   {
     return Optional.empty();
   }
